+++ examples/nextjs/next.config.js
@@ -1,7 +1,11 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
-  transpilePackages: ['@btc-connect/react', '@btc-connect/ui', '@btc-connect/core'],
+  transpilePackages: [
+    '@btc-connect/react',
+    '@btc-connect/ui',
+    '@btc-connect/core',
+  ],
   webpack: (config) => {
     // Support for Web Components
     config.experiments = {