+++ examples/vue-example/src/main.ts
@@ -1,11 +1,11 @@
-import { createApp } from 'vue'
-import './style.css'
-import App from './App.vue'
+import { createApp } from 'vue';
+import './style.css';
 
-import { BTCWalletPlugin } from '@btc-connect/vue'
+import { BTCWalletPlugin } from '@btc-connect/vue';
+import App from './App.vue';
 
-const app = createApp(App)
+const app = createApp(App);
 app.use(BTCWalletPlugin, {
   autoConnect: true,
-})
-app.mount('#app')
+});
+app.mount('#app');